REF: .get tests (#33164)
